TEST: finish first round of smolyak tests
There are still a couple things that could be improved, these have
been noted as TODO items in test/smol.jl

+++ src/smolyak.jl
@@ -112,7 +112,7 @@ function Base.show(io::IO, p::SmolyakParams)
 end
 
 # TODO: fix this
-function Base.length{T,Ti<:Integer}(sp::SmolyakParams{T,Ti})
+function Base.length{T,Ti<:Integer}(sp::SmolyakParams{T,Ti})::Int
     d, mu = sp.d, sp.mu
     mu == 1 ? 2d - 1 :
     mu == 2 ? Int(1 + 4d + 4d*(d-1)/2 ):
